Refactor-Entities
Moví las entidades de las carpetas de sus respectivos modulos a una carpeta de entidades
