+++ src/groups/groups.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Group } from './group.entity';
+import { Group } from '../entities/group.entity';
 
 @Module({imports: [TypeOrmModule.forFeature([Group])],})
 export class GroupsModule {}
