+++ src/tasks/tasks.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Task } from './task.entity';
+import { Task } from '../entities/task.entity';
 
 @Module({imports: [TypeOrmModule.forFeature([Task])],})
 export class TasksModule {}
