+++ src/users/users.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from './user.entity';
-import { Task } from 'src/tasks/task.entity';
-import { Group } from 'src/groups/group.entity';
+import { User } from '../entities/user.entity';
+import { Task } from 'src/entities/task.entity';
+import { Group } from 'src/entities/group.entity';
 
 @Module({imports: [TypeOrmModule.forFeature([User])],})
 export class UsersModule {}
